Move and refactor test as now the UI is not anymore
